# docs/openapi.yaml
openapi: 3.0.3
info:
  title: A2A API Gateway
  description: |
    نظام تكامل شامل للتطبيقات (Application-to-Application) يربط بين التطبيقات المحلية ونظام الأوتوبيلوت وبوت تيليجرام.
    
    ## الميزات الرئيسية
    - API Gateway موحد لإدارة جميع الاتصالات
    - Message Broker للتواصل غير المتزامن
    - تكامل الأوتوبيلوت لإدارة المهام التلقائية
    - تكامل تيليجرام للتفاعل مع المستخدمين
    - نظام مراقبة شامل مع Prometheus و Grafana
    - أمان متقدم مع JWT و Rate Limiting
  version: 1.0.0
  contact:
    name: AuraOS Team
    email: support@auraos.com
    url: https://auraos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.auraos.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and get access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: string
                    example: 24h
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register:
    post:
      summary: User registration
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: newuser
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                role:
                  type: string
                  enum: [user, admin, service]
                  example: user
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: User created successfully
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/refresh:
    post:
      summary: Refresh token
      description: Refresh authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expiresIn:
                    type: string
                    example: 24h
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for service authentication
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user_1234567890_abc123
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@auraos.com
        role:
          type: string
          enum: [user, admin, service]
          example: admin
        permissions:
          type: array
          items:
            type: string
          example: [read, write, admin]
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Invalid request parameters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: username
              message:
                type: string
                example: Username is required
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: req_1234567890_abc123
...